refactor(nuget): split client.go by concern and flatten GetDeprecation
Split the 799-line nuget/client.go into four focused files:
- client.go: Client struct, NewClient, Set* methods, cacheEntry, remember
- deprecation.go: DeprecationInfo, GetDeprecation, registration* wire types,
  extractFirstDeprecation, scrapeDeprecationFromNuGetHTML
- repo_url.go: GetRepoURL, extractRepoURLFromLeaves, normalizeRepoURL,
  resolveRepoURLHeuristics, followRedirect, scrapeFirstGitHubFromHTML
- service_index.go: ServiceIndex, getRegistrationCandidates, discoverRegistrationBases

Also extract the inner page-fetch loop from the 128-line GetDeprecation into
deprecationFromIndex(ctx, reg, id) to flatten the nested control flow.
Types are placed before their first user per project convention.
