+++ internal/infrastructure/nuget/client.go
@@ -1,20 +1,11 @@
 package nuget
 
 import (
-	"context"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
-	"log/slog"
 	"net/http"
-	"net/url"
-	"regexp"
 	"strings"
 	"sync"
 	"time"
 
-	"github.com/future-architect/uzomuzo-oss/internal/common"
 	"github.com/future-architect/uzomuzo-oss/internal/infrastructure/httpclient"
 )
 
@@ -128,152 +119,14 @@ func (c *Client) SetHTMLBase(u string) {
 	c.mu.Unlock()
 }
 
-// DeprecationInfo carries NuGet deprecation metadata (if any).
-type DeprecationInfo struct {
-	Reasons            []string
-	Message            string
-	AlternatePackageID string
-}
-
+// cacheEntry stores a deprecation lookup result with its retrieval timestamp.
 type cacheEntry struct {
 	info      *DeprecationInfo
 	found     bool
 	fetchedAt time.Time
 }
 
-// GetDeprecation fetches deprecation information for a package id.
-// Returns (info, found, error). When found=false, info is nil.
-func (c *Client) GetDeprecation(ctx context.Context, packageID string) (*DeprecationInfo, bool, error) {
-	id := strings.TrimSpace(packageID)
-	if id == "" {
-		return nil, false, fmt.Errorf("package id is required")
-	}
-
-	// Cache
-	if c.cacheTTL > 0 {
-		c.mu.Lock()
-		if ce, ok := c.cache[id]; ok {
-			if time.Since(ce.fetchedAt) < c.cacheTTL {
-				info := ce.info
-				found := ce.found
-				c.mu.Unlock()
-				slog.Debug("nuget: cache hit for deprecation", "id", id, "found", found)
-				return info, found, nil
-			}
-		}
-		c.mu.Unlock()
-	}
-
-	// Registration index endpoint (with fallback base variants)
-	idLower := strings.ToLower(id) // NuGet registration path requires lowercase package ID
-	candidates := c.getRegistrationCandidates(ctx)
-
-	var last404 bool
-	var anyOK bool
-	var decodeErrors int
-	for idx, b := range candidates {
-		endpoint := fmt.Sprintf("%s/%s/index.json", b, url.PathEscape(idLower))
-		slog.Debug("nuget: request registration index", "id", id, "endpoint", endpoint, "attempt", idx+1)
-		req, err := http.NewRequestWithContext(ctx, http.MethodGet, endpoint, nil)
-		if err != nil {
-			return nil, false, fmt.Errorf("failed to build NuGet request: %w", err)
-		}
-		// Be a good citizen: set a descriptive User-Agent
-		req.Header.Set("User-Agent", "uzomuzo-nuget-client/1.0 (+https://github.com/future-architect/uzomuzo-oss)")
-		resp, err := c.http.Do(ctx, req)
-		if err != nil {
-			return nil, false, fmt.Errorf("NuGet HTTP error: %w", err)
-		}
-		if resp.StatusCode == http.StatusNotFound {
-			slog.Debug("nuget: registration index not found", "id", id, "endpoint", endpoint)
-			last404 = true
-			_ = resp.Body.Close() // best-effort cleanup
-			continue
-		}
-		if resp.StatusCode != http.StatusOK {
-			slog.Debug("nuget: non-OK status from registration index", "id", id, "status", resp.StatusCode, "endpoint", endpoint)
-			_ = resp.Body.Close() // best-effort cleanup
-			// Non-OK other than 404: treat as error for now
-			return nil, false, fmt.Errorf("NuGet HTTP %d", resp.StatusCode)
-		}
-
-		anyOK = true
-		var reg registrationIndex
-		if err := json.NewDecoder(resp.Body).Decode(&reg); err != nil {
-			// Do not abort on decode errors (e.g., slow network/timeouts while reading body).
-			// Try next candidate (gz/non-gz) or fallback scraper.
-			slog.Debug("nuget: decode failed for registration index", "id", id, "error", err)
-			_ = resp.Body.Close() // best-effort cleanup
-			decodeErrors++
-			continue
-		}
-		_ = resp.Body.Close() // best-effort cleanup
-
-		// Iterate pages; items may be embedded or referenced.
-		for _, page := range reg.Items {
-			// Embedded leaves
-			if len(page.Items) > 0 {
-				if info, ok := extractFirstDeprecation(page.Items); ok {
-					slog.Debug("nuget: deprecation found (embedded)", "id", id, "reasons", info.Reasons, "alt", info.AlternatePackageID)
-					c.remember(id, info)
-					return info, true, nil
-				}
-				continue
-			}
-			// Need to fetch page
-			if page.ID == "" {
-				continue
-			}
-			slog.Debug("nuget: fetching registration page", "id", id, "page", page.ID)
-			piReq, err := http.NewRequestWithContext(ctx, http.MethodGet, page.ID, nil)
-			if err != nil {
-				return nil, false, fmt.Errorf("NuGet request (page) failed: %w", err)
-			}
-			piReq.Header.Set("User-Agent", "uzomuzo-nuget-client/1.0 (+https://github.com/future-architect/uzomuzo-oss)")
-			piResp, err := c.http.Do(ctx, piReq)
-			if err != nil {
-				return nil, false, fmt.Errorf("NuGet HTTP (page) error: %w", err)
-			}
-			if piResp.StatusCode != http.StatusOK {
-				_ = piResp.Body.Close() // best-effort cleanup
-				slog.Debug("nuget: non-OK page status", "id", id, "status", piResp.StatusCode, "page", page.ID)
-				continue
-			}
-			var leaf registrationPage
-			if err := json.NewDecoder(piResp.Body).Decode(&leaf); err != nil {
-				_ = piResp.Body.Close() // best-effort cleanup
-				slog.Debug("nuget: page decode failed", "id", id, "error", err)
-				return nil, false, fmt.Errorf("NuGet page decode failed: %w", err)
-			}
-			_ = piResp.Body.Close() // best-effort cleanup
-			if info, ok := extractFirstDeprecation(leaf.Items); ok {
-				slog.Debug("nuget: deprecation found (page)", "id", id, "reasons", info.Reasons, "alt", info.AlternatePackageID)
-				c.remember(id, info)
-				return info, true, nil
-			}
-		}
-		// No deprecation in this variant, try next
-		slog.Debug("nuget: no deprecation in this index variant", "id", id, "endpoint", endpoint)
-	}
-
-	// No deprecation found across all attempts
-	if !anyOK && last404 {
-		slog.Debug("nuget: no deprecation found (all variants 404)", "id", id)
-	} else {
-		slog.Debug("nuget: no deprecation found (checked variants)", "id", id)
-	}
-	// Best-effort HTML fallback for nuget.org: scrape the package page for a deprecation banner
-	if info, ok := c.scrapeDeprecationFromNuGetHTML(ctx, id); ok {
-		slog.Debug("nuget: deprecation found via HTML fallback", "id", id, "alt", info.AlternatePackageID)
-		c.remember(id, info)
-		return info, true, nil
-	}
-	c.remember(id, nil)
-	return nil, false, nil
-}
-
-// fetchLeafDeprecation removed: leaf-by-leaf crawling is unnecessary after embedded/page extraction handles deprecation.
-
+// remember stores a deprecation lookup result in the cache, respecting NoCacheNotFound.
 func (c *Client) remember(id string, info *DeprecationInfo) {
 	if c.cacheTTL <= 0 {
 		return
@@ -286,514 +139,4 @@ func (c *Client) remember(id string, info *DeprecationInfo) {
 	}
 	c.cache[id] = cacheEntry{info: info, found: info != nil, fetchedAt: time.Now()}
 	c.mu.Unlock()
-}
-
-// Minimal JSON shapes for Registration API
-type registrationIndex struct {
-	Items []registrationPageRef `json:"items"`
-}
-
-type registrationPageRef struct {
-	ID    string             `json:"@id"`
-	Items []registrationLeaf `json:"items"`
-}
-
-type registrationPage struct {
-	Items []registrationLeaf `json:"items"`
-}
-
-type registrationLeaf struct {
-	ID string `json:"@id"`
-	// CatalogEntry can be either an object or a string URL depending on embedding.
-	// We don't need its contents for deprecation checks, so accept any JSON.
-	CatalogEntry json.RawMessage          `json:"catalogEntry"`
-	Deprecation  *registrationDeprecation `json:"deprecation"`
-}
-
-type registrationDeprecation struct {
-	Reasons          []string                `json:"reasons"`
-	Message          string                  `json:"message"`
-	AlternatePackage *registrationAltPackage `json:"alternatePackage"`
-}
-
-type registrationAltPackage struct {
-	ID    string `json:"id"`
-	Range string `json:"range"`
-}
-
-// catalogEntryDoc represents the minimal shape of a catalog entry when embedded or fetched.
-type catalogEntryDoc struct {
-	Deprecation *registrationDeprecation `json:"deprecation"`
-}
-
-func extractFirstDeprecation(items []registrationLeaf) (*DeprecationInfo, bool) {
-	for _, it := range items {
-		// 1) Leaf-level deprecation
-		if it.Deprecation != nil {
-			altID := ""
-			if it.Deprecation.AlternatePackage != nil {
-				altID = strings.TrimSpace(it.Deprecation.AlternatePackage.ID)
-			}
-			info := &DeprecationInfo{
-				Reasons:            append([]string(nil), it.Deprecation.Reasons...),
-				Message:            strings.TrimSpace(it.Deprecation.Message),
-				AlternatePackageID: altID,
-			}
-			return info, true
-		}
-		// 2) catalogEntry may be embedded as an object and include deprecation
-		if len(it.CatalogEntry) > 0 && len(it.CatalogEntry) >= 1 && it.CatalogEntry[0] == '{' {
-			var ce catalogEntryDoc
-			if err := json.Unmarshal(it.CatalogEntry, &ce); err == nil && ce.Deprecation != nil {
-				altID := ""
-				if ce.Deprecation.AlternatePackage != nil {
-					altID = strings.TrimSpace(ce.Deprecation.AlternatePackage.ID)
-				}
-				info := &DeprecationInfo{
-					Reasons:            append([]string(nil), ce.Deprecation.Reasons...),
-					Message:            strings.TrimSpace(ce.Deprecation.Message),
-					AlternatePackageID: altID,
-				}
-				return info, true
-			}
-		}
-	}
-	return nil, false
-}
-
-// GetRepoURL attempts to extract a repository URL (preferably GitHub) for the given NuGet package ID.
-//
-// DDD Layer: Infrastructure
-// Behavior:
-//   - Queries the NuGet Registration index like GetDeprecation
-//   - Scans embedded page leaves for an embedded catalogEntry and extracts repository/project URL fields
-//   - If needed, fetches registration pages to inspect embedded catalogEntry entries
-//   - Returns a normalized URL or an empty string when not determinable
-func (c *Client) GetRepoURL(ctx context.Context, packageID string, _ string) (string, error) {
-	id := strings.TrimSpace(packageID)
-	if id == "" {
-		return "", fmt.Errorf("package id is required")
-	}
-
-	idLower := strings.ToLower(id)
-	candidates := c.getRegistrationCandidates(ctx)
-
-	for idx, b := range candidates {
-		endpoint := fmt.Sprintf("%s/%s/index.json", b, url.PathEscape(idLower))
-		slog.Debug("nuget: request registration index (repo)", "id", id, "endpoint", endpoint, "attempt", idx+1)
-		req, err := http.NewRequestWithContext(ctx, http.MethodGet, endpoint, nil)
-		if err != nil {
-			return "", fmt.Errorf("failed to build NuGet request: %w", err)
-		}
-		req.Header.Set("User-Agent", "uzomuzo-nuget-client/1.0 (+https://github.com/future-architect/uzomuzo-oss)")
-		resp, err := c.http.Do(ctx, req)
-		if err != nil {
-			return "", fmt.Errorf("NuGet HTTP error: %w", err)
-		}
-		if resp.StatusCode == http.StatusNotFound {
-			_ = resp.Body.Close() // best-effort cleanup
-			continue
-		}
-		if resp.StatusCode != http.StatusOK {
-			_ = resp.Body.Close() // best-effort cleanup
-			return "", fmt.Errorf("NuGet HTTP %d", resp.StatusCode)
-		}
-		var reg registrationIndex
-		if err := json.NewDecoder(resp.Body).Decode(&reg); err != nil {
-			_ = resp.Body.Close() // best-effort cleanup
-			return "", fmt.Errorf("NuGet decode failed: %w", err)
-		}
-		_ = resp.Body.Close() // best-effort cleanup
-
-		// 1) Inspect embedded leaves for catalogEntry object
-		for _, page := range reg.Items {
-			if len(page.Items) > 0 {
-				if repo := extractRepoURLFromLeaves(page.Items); repo != "" {
-					if resolved := c.resolveRepoURLHeuristics(ctx, repo); resolved != "" {
-						return resolved, nil
-					}
-					return repo, nil
-				}
-				continue
-			}
-			if page.ID == "" {
-				continue
-			}
-			// 2) Fetch page document and inspect its items
-			piReq, err := http.NewRequestWithContext(ctx, http.MethodGet, page.ID, nil)
-			if err != nil {
-				return "", fmt.Errorf("NuGet request (page) failed: %w", err)
-			}
-			piReq.Header.Set("User-Agent", "uzomuzo-nuget-client/1.0 (+https://github.com/future-architect/uzomuzo-oss)")
-			piResp, err := c.http.Do(ctx, piReq)
-			if err != nil {
-				return "", fmt.Errorf("NuGet HTTP (page) error: %w", err)
-			}
-			if piResp.StatusCode != http.StatusOK {
-				_ = piResp.Body.Close() // best-effort cleanup
-				continue
-			}
-			var leaf registrationPage
-			if err := json.NewDecoder(piResp.Body).Decode(&leaf); err != nil {
-				_ = piResp.Body.Close() // best-effort cleanup
-				return "", fmt.Errorf("NuGet page decode failed: %w", err)
-			}
-			_ = piResp.Body.Close() // best-effort cleanup
-			if repo := extractRepoURLFromLeaves(leaf.Items); repo != "" {
-				if resolved := c.resolveRepoURLHeuristics(ctx, repo); resolved != "" {
-					return resolved, nil
-				}
-				return repo, nil
-			}
-		}
-	}
-	return "", nil
-}
-
-// extractRepoURLFromLeaves inspects registration leaves for an embedded catalogEntry object
-// and attempts to extract a repository or project URL from it. Returns a normalized URL or empty string.
-func extractRepoURLFromLeaves(items []registrationLeaf) string {
-	for _, it := range items {
-		if len(it.CatalogEntry) == 0 || it.CatalogEntry[0] != '{' {
-			continue
-		}
-		// Use a permissive map to handle variations: repository (object or string), projectUrl, repositoryUrl
-		var m map[string]any
-		if err := json.Unmarshal(it.CatalogEntry, &m); err != nil {
-			continue
-		}
-		// Try repository (object with url)
-		if v, ok := m["repository"]; ok {
-			switch rv := v.(type) {
-			case map[string]any:
-				if u, ok := rv["url"].(string); ok && strings.TrimSpace(u) != "" {
-					if norm := normalizeRepoURL(u); norm != "" {
-						return norm
-					}
-				}
-			case string:
-				if norm := normalizeRepoURL(rv); norm != "" {
-					return norm
-				}
-			}
-		}
-		// Try repositoryUrl (string)
-		if u, ok := m["repositoryUrl"].(string); ok && strings.TrimSpace(u) != "" {
-			if norm := normalizeRepoURL(u); norm != "" {
-				return norm
-			}
-		}
-		// Fallback to projectUrl (string)
-		if u, ok := m["projectUrl"].(string); ok && strings.TrimSpace(u) != "" {
-			if norm := normalizeRepoURL(u); norm != "" {
-				return norm
-			}
-		}
-	}
-	return ""
-}
-
-// normalizeRepoURL trims and returns the URL as-is; the consumer will further normalize
-// to GitHub project keys if needed. Keep it minimal here.
-func normalizeRepoURL(s string) string {
-	return strings.TrimSpace(s)
-}
-
-// getRegistrationCandidates returns a list of Registration Base URLs to try.
-// Order of precedence:
-//  1. Discovered bases from the service index (cached with TTL)
-//  2. Configured baseURL
-//  3. Sibling variant on nuget.org (semver2 <-> gz-semver2)
-//  4. Fallback to known nuget.org gz-semver2 base
-func (c *Client) getRegistrationCandidates(ctx context.Context) []string {
-	base := strings.TrimRight(c.baseURL, "/")
-	// If baseURL is overridden (not the default), honor it and skip network discovery
-	if base != strings.TrimRight(defaultRegistrationBase, "/") {
-		candidates := []string{base}
-		if strings.Contains(base, "registration5-semver2") && !strings.Contains(base, "registration5-gz-semver2") {
-			candidates = append(candidates, strings.Replace(base, "registration5-semver2", "registration5-gz-semver2", 1))
-		} else if strings.Contains(base, "registration5-gz-semver2") {
-			candidates = append(candidates, strings.Replace(base, "registration5-gz-semver2", "registration5-semver2", 1))
-		} else {
-			candidates = append(candidates, "https://api.nuget.org/v3/registration5-gz-semver2")
-		}
-		return candidates
-	}
-
-	// Otherwise, try discovery first
-	discovered := c.discoverRegistrationBases(ctx)
-	if len(discovered) > 0 {
-		return discovered
-	}
-	candidates := []string{base}
-	if strings.Contains(base, "registration5-semver2") && !strings.Contains(base, "registration5-gz-semver2") {
-		candidates = append(candidates, strings.Replace(base, "registration5-semver2", "registration5-gz-semver2", 1))
-	} else if strings.Contains(base, "registration5-gz-semver2") {
-		candidates = append(candidates, strings.Replace(base, "registration5-gz-semver2", "registration5-semver2", 1))
-	} else {
-		candidates = append(candidates, "https://api.nuget.org/v3/registration5-gz-semver2")
-	}
-	return candidates
-}
-
-// ServiceIndex represents the minimal shape of the NuGet v3 service index.
-type ServiceIndex struct {
-	Resources []struct {
-		ID   string `json:"@id"`
-		Type string `json:"@type"`
-	} `json:"resources"`
-}
-
-// discoverRegistrationBases loads the service index (if TTL expired) and extracts Registration Base URLs.
-// It caches the discovered list and returns it for use by callers.
-func (c *Client) discoverRegistrationBases(ctx context.Context) []string {
-	c.mu.Lock()
-	idxURL := c.serviceIndexURL
-	ttl := c.serviceIndexTTL
-	// If cache is fresh, return it immediately.
-	if len(c.discoveredBases) > 0 && time.Since(c.discoveredAt) < ttl {
-		bases := append([]string(nil), c.discoveredBases...)
-		c.mu.Unlock()
-		return bases
-	}
-	c.mu.Unlock()
-
-	// No fresh cache, attempt discovery.
-	if strings.TrimSpace(idxURL) == "" {
-		return nil
-	}
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, idxURL, nil)
-	if err != nil {
-		return nil
-	}
-	req.Header.Set("User-Agent", "uzomuzo-nuget-client/1.0 (+https://github.com/future-architect/uzomuzo-oss)")
-	resp, err := c.http.Do(ctx, req)
-	if err != nil {
-		return nil
-	}
-	defer func() { _ = resp.Body.Close() }()
-	if resp.StatusCode != http.StatusOK {
-		return nil
-	}
-	var idx ServiceIndex
-	if err := json.NewDecoder(resp.Body).Decode(&idx); err != nil {
-		return nil
-	}
-
-	// Collect registration base URLs. Prefer semver2, then gz-semver2, then any RegistrationsBaseUrl.
-	var semver2 []string
-	var gzSemver2 []string
-	var generic []string
-	for _, r := range idx.Resources {
-		t := strings.ToLower(r.Type)
-		if r.ID == "" || t == "" {
-			continue
-		}
-		switch {
-		case strings.Contains(t, "registrationsbaseurl/3.6.0") || strings.Contains(t, "registrationsbaseurl/3.5.0") || strings.Contains(t, "registrationsbaseurl/3.0.0"):
-			// Generic match; we will still sort by content of the URL for semver2 preference later
-			if strings.Contains(strings.ToLower(r.ID), "registration5-semver2") {
-				semver2 = append(semver2, strings.TrimRight(r.ID, "/"))
-			} else if strings.Contains(strings.ToLower(r.ID), "registration5-gz-semver2") {
-				gzSemver2 = append(gzSemver2, strings.TrimRight(r.ID, "/"))
-			} else {
-				generic = append(generic, strings.TrimRight(r.ID, "/"))
-			}
-		case strings.Contains(t, "registrationsbaseurl"):
-			// Fallback broad match
-			if strings.Contains(strings.ToLower(r.ID), "registration5-semver2") {
-				semver2 = append(semver2, strings.TrimRight(r.ID, "/"))
-			} else if strings.Contains(strings.ToLower(r.ID), "registration5-gz-semver2") {
-				gzSemver2 = append(gzSemver2, strings.TrimRight(r.ID, "/"))
-			} else {
-				generic = append(generic, strings.TrimRight(r.ID, "/"))
-			}
-		}
-	}
-	// Merge with precedence: semver2 first, then gz-semver2, then generic.
-	dedup := make(map[string]bool)
-	var bases []string
-	for _, list := range [][]string{semver2, gzSemver2, generic} {
-		for _, b := range list {
-			if !dedup[b] {
-				dedup[b] = true
-				bases = append(bases, b)
-			}
-		}
-	}
-	if len(bases) == 0 {
-		return nil
-	}
-	c.mu.Lock()
-	c.discoveredBases = bases
-	c.discoveredAt = time.Now()
-	c.mu.Unlock()
-	return append([]string(nil), bases...)
-}
-
-// resolveRepoURLHeuristics attempts to improve a non-GitHub repository URL returned by NuGet metadata.
-// For Microsoft packages, NuGet often provides go.microsoft.com/aka.ms short links or docs pages.
-// This function will try to follow redirects and, if landing on a docs page, scrape a GitHub repo URL.
-func (c *Client) resolveRepoURLHeuristics(ctx context.Context, raw string) string {
-	s := strings.TrimSpace(raw)
-	if s == "" {
-		return ""
-	}
-	u, err := url.Parse(s)
-	if err != nil || u.Hostname() == "" {
-		return ""
-	}
-	host := strings.ToLower(u.Hostname())
-
-	// Only attempt network heuristics for Microsoft shorteners/docs
-	if host == "aka.ms" || strings.HasSuffix(host, ".microsoft.com") {
-		if final := c.followRedirect(ctx, s); final != "" {
-			fu, _ := url.Parse(final)
-			if fu != nil {
-				if strings.Contains(strings.ToLower(fu.Hostname()), "github.com") {
-					// Normalize and validate before returning
-					if norm := common.NormalizeRepositoryURL(final); norm != "" && common.IsValidGitHubURL(norm) {
-						return norm
-					}
-					if common.IsValidGitHubURL(final) {
-						return final
-					}
-				}
-				// If we landed on a docs page, attempt to find a GitHub link within the HTML
-				if strings.Contains(strings.ToLower(fu.Hostname()), "docs.microsoft.com") || strings.Contains(strings.ToLower(fu.Hostname()), "learn.microsoft.com") {
-					if gh := c.scrapeFirstGitHubFromHTML(ctx, final); gh != "" {
-						return gh
-					}
-				}
-			}
-		}
-	}
-	return ""
-}
-
-// followRedirect performs a GET request and returns the final URL after redirects.
-func (c *Client) followRedirect(ctx context.Context, startURL string) string {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, startURL, nil)
-	if err != nil {
-		return ""
-	}
-	req.Header.Set("User-Agent", "uzomuzo-nuget-client/1.0 (+https://github.com/future-architect/uzomuzo-oss)")
-	resp, err := c.http.Do(ctx, req)
-	if err != nil {
-		return ""
-	}
-	defer func() { _ = resp.Body.Close() }()
-	// Drain small body to allow re-use
-	_, _ = io.CopyN(io.Discard, resp.Body, 1024) // best-effort drain before close
-	if resp.Request != nil && resp.Request.URL != nil {
-		return resp.Request.URL.String()
-	}
-	return ""
-}
-
-// scrapeFirstGitHubFromHTML fetches the page and returns the first GitHub repository URL found in the HTML.
-func (c *Client) scrapeFirstGitHubFromHTML(ctx context.Context, pageURL string) string {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, pageURL, nil)
-	if err != nil {
-		return ""
-	}
-	req.Header.Set("User-Agent", "uzomuzo-nuget-client/1.0 (+https://github.com/future-architect/uzomuzo-oss)")
-	resp, err := c.http.Do(ctx, req)
-	if err != nil {
-		return ""
-	}
-	defer func() { _ = resp.Body.Close() }()
-	if resp.StatusCode < 200 || resp.StatusCode >= 400 {
-		return ""
-	}
-	// Read up to 512KB to avoid large downloads
-	const maxRead = 512 * 1024
-	limited := io.LimitReader(resp.Body, maxRead)
-	body, err := io.ReadAll(limited)
-	if err != nil && !errors.Is(err, io.EOF) {
-		return ""
-	}
-	// Simple regex to find a GitHub URL
-	re := regexp.MustCompile(`https?://github\.com/[^"'\s<>]+`)
-	if m := re.Find(body); len(m) > 0 {
-		// Normalize candidate URL and validate it's a GitHub repo URL
-		candidate := string(m)
-		normalized := common.NormalizeRepositoryURL(candidate)
-		if normalized != "" && common.IsValidGitHubURL(normalized) {
-			return normalized
-		}
-		// Fallback to raw candidate if normalization empties, but still looks valid
-		if common.IsValidGitHubURL(candidate) {
-			return candidate
-		}
-	}
-	return ""
-}
-
-// scrapeDeprecationFromNuGetHTML performs a best-effort parse of the nuget.org package page
-// to determine if the package is deprecated and to extract a suggested successor.
-// It returns (info, true) when a deprecation banner is detected. Message and successor are optional.
-func (c *Client) scrapeDeprecationFromNuGetHTML(ctx context.Context, id string) (*DeprecationInfo, bool) {
-	base := c.htmlBase
-	if strings.TrimSpace(base) == "" {
-		base = "https://www.nuget.org"
-	}
-	// Use the provided ID as-is; nuget.org routing is case-insensitive.
-	pageURL := fmt.Sprintf("%s/packages/%s", strings.TrimRight(base, "/"), url.PathEscape(id))
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, pageURL, nil)
-	if err != nil {
-		return nil, false
-	}
-	req.Header.Set("User-Agent", "uzomuzo-nuget-client/1.0 (+https://github.com/future-architect/uzomuzo-oss)")
-	resp, err := c.http.Do(ctx, req)
-	if err != nil {
-		return nil, false
-	}
-	defer func() { _ = resp.Body.Close() }()
-	if resp.StatusCode < 200 || resp.StatusCode >= 400 {
-		return nil, false
-	}
-	// Read up to 512KB
-	const maxRead = 512 * 1024
-	body, err := io.ReadAll(io.LimitReader(resp.Body, maxRead))
-	if err != nil && !errors.Is(err, io.EOF) {
-		return nil, false
-	}
-	lower := strings.ToLower(string(body))
-	if !strings.Contains(lower, "this package has been deprecated") && !strings.Contains(lower, "deprecated") {
-		return nil, false
-	}
-	// Try to extract a successor from a Suggested Alternatives section link
-	// Example snippet:
-	//   Suggested Alternatives
-	//   <a href="/packages/Azure.Messaging.EventHubs">Azure.Messaging.EventHubs</a>
-	reAlt := regexp.MustCompile(`(?is)suggested\s+alternatives?.{0,400}?<a[^>]*href="/packages/([A-Za-z0-9_.\-]+)/?"[^>]*>\s*([^<]+)\s*</a>`)
-	var altID string
-	if m := reAlt.FindSubmatch(body); len(m) >= 2 {
-		altID = strings.TrimSpace(string(m[1]))
-	}
-	// Extract a brief message from the deprecation area if present
-	// Heuristic: grab up to 200 chars around the first occurrence of "deprecated"
-	msg := ""
-	if idx := strings.Index(lower, "deprecated"); idx >= 0 {
-		start := idx - 80
-		if start < 0 {
-			start = 0
-		}
-		end := idx + 200
-		if end > len(body) {
-			end = len(body)
-		}
-		snippet := string(body[start:end])
-		// Collapse whitespace
-		snippet = regexp.MustCompile(`\s+`).ReplaceAllString(snippet, " ")
-		msg = strings.TrimSpace(snippet)
-		if len(msg) > 300 {
-			msg = msg[:300] + "…"
-		}
-	}
-	// We do not know the NuGet machine reason (Legacy/CriticalBugs/Other) from HTML alone.
-	// Provide a neutral reason to indicate deprecation was detected via HTML.
-	reasons := []string{"Other"}
-	return &DeprecationInfo{Reasons: reasons, Message: msg, AlternatePackageID: altID}, true
 }
